test(#1267): split issue_992 ttl/tombstone/range parity test by scenario (under ratchet)
